UI text is now put into a central registry
I have no idea if this is a good idea or not, but I want any user
interface facing strings to be owned by a singleton, basically.
This is because un-atlasable strings require a texture to show on
screen, and I don't want a bunch of small objects having and releasing
ownership of textures all the time.

* UIStringPool: set of strings, map of strings to textures
* UI strings are just views, which can index into the texture map
* Registering a string with the pool gives you a string_view (the map key)

* Implemented a collections (e.g. albums, playlists) table because
  of my squirrel brain

* Removed DynamicText class because Textures already have all that info
* Improved constness of reference parameters in various places
* Improved some naming and namespacing of text utils